+++ Scripts/setup_dependencies.sh
@@ -21,6 +21,24 @@ fi
 brew update
 # check if "docker" in dependencies
 if [[ $dependencies == *"docker"* ]]; then
-    brew install --cask docker
+    brew install --cask docker || echo "Install Docker manually"
 fi
+
 brew install $dependencies
+
+if [[ $dependencies == *"pyenv"* ]]; then
+    # If pyenv not already exists in zshrc, add it
+    if [ -f ~/.zshrc ] && [ -x "$(command -v pyenv)" ]; then
+        echo 'export PYENV_ROOT="$HOME/.pyenv"' >> ~/.zshrc
+        echo 'command -v pyenv >/dev/null || export PATH="$PYENV_ROOT/bin:$PATH"' >> ~/.zshrc
+        echo 'eval "$(pyenv init -)"' >> ~/.zshrc
+    fi
+
+    export PYENV_ROOT="$HOME/.pyenv";
+    export PATH="$PYENV_ROOT/bin":$PATH;
+    eval "$(pyenv init -)" || echo "No pyenv, procceding without";
+    # get latest python version from the pyenv list
+    LATEST_PYTHON=$(pyenv install --list | grep --extended-regexp "^\s*[0-9][0-9.]*[0-9]\s*$" | tail -1);
+    pyenv install "$LATEST_PYTHON" 2.7.18;
+    pyenv global "$LATEST_PYTHON" 2.7.18;
+fi